adapted tests
